+++ src/screens/ExchangeScreen/SearchTokens/SearchTokens.styles.ts
