+++ src/screens/ExchangeScreen/TokenCard/TokenCard.types.ts
@@ -1,0 +1,13 @@
+import { RefObject } from 'react';
+import { TextInput } from 'react-native';
+import { ParaswapToken } from '@models/token';
+
+export interface TokenCardProps {
+	token: ParaswapToken | undefined;
+	onPress?: (() => void) | undefined;
+	balance: string;
+	innerRef?: RefObject<TextInput>;
+	disableMax?: boolean;
+	updateQuotes?: Function;
+	conversionAmount?: string;
+}
