+++ src/screens/ExchangeScreen/Warning/Warning.styles.ts
@@ -1,0 +1,23 @@
+import { StyleSheet } from 'react-native';
+
+export const styles = StyleSheet.create({
+	container: {
+		flexDirection: 'row',
+		borderRadius: 16,
+		marginBottom: 32,
+		borderColor: '#FFFCF5',
+		borderWidth: 1
+	},
+	iconContainer: {
+		backgroundColor: '#FFFCF5',
+		paddingVertical: 20,
+		paddingHorizontal: 12,
+		borderTopLeftRadius: 16,
+		borderBottomLeftRadius: 16
+	},
+	text: {
+		paddingVertical: 10,
+		paddingHorizontal: 16,
+		alignSelf: 'center'
+	}
+});
