+++ src/screens/ExchangeScreen/Warning/Warning.tsx
@@ -1,29 +1,14 @@
 import React from 'react';
 import { View } from 'react-native';
 import { Icon, Text } from '@components';
+import { styles } from './Warning.styles';
 
 const Warning = ({ label }: { label: string }) => (
-	<View
-		style={{
-			flexDirection: 'row',
-			borderRadius: 16,
-			marginBottom: 32,
-			borderColor: '#FFFCF5',
-			borderWidth: 1
-		}}
-	>
-		<View
-			style={{
-				backgroundColor: '#FFFCF5',
-				paddingVertical: 20,
-				paddingHorizontal: 12,
-				borderTopLeftRadius: 16,
-				borderBottomLeftRadius: 16
-			}}
-		>
+	<View style={styles.container}>
+		<View style={styles.iconContainer}>
 			<Icon name="alertStroke" color="alert4a" size={32} />
 		</View>
-		<Text type="a" style={{ paddingVertical: 10, paddingHorizontal: 16, alignSelf: 'center' }}>
+		<Text type="a" style={styles.text}>
 			{label}
 		</Text>
 	</View>
